adapter: fix off-by-one in connection id generation
We were 1 bit high on the max connection id which ended up conflicting
with the mask.

Fixes #24574
