+++ src/balancerd/src/lib.rs
@@ -29,9 +29,9 @@ use axum::{routing, Router};
 use bytes::BytesMut;
 use futures::TryFutureExt;
 use hyper::StatusCode;
-use mz_adapter_types::connection::conn_id_org_uuid;
 use mz_build_info::{build_info, BuildInfo};
 use mz_frontegg_auth::Authentication as FronteggAuthentication;
+use mz_ore::id_gen::conn_id_org_uuid;
 use mz_ore::metrics::{ComputedGauge, IntCounter, IntGauge, MetricsRegistry};
 use mz_ore::netio::AsyncReady;
 use mz_ore::task::{spawn, JoinSetExt};
